# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Derive API",
    "version": "1.0.0",
    "description": "Derive's HTTP API. Every response shape here is the single source of truth for the web client — generated from the same Zod schemas that back this reference, so the docs never drift from the running API.\n\n**Auth:** a session cookie (browser) or a bearer token (agents and the CLI). Most write routes are scoped to the caller's active workspace. The `/v1` paths are the stable product surface; `/openapi.json` serves this spec and `/docs` renders it."
  },
...
  "components": {
    "schemas": {
...
      "Viewer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The viewer's id: a signed-in user's id, or a stable anonymous viewer id"
          },
          "name": {
            "type": "string",
            "description": "Display name: a user's @handle, or a friendly handle for anonymous viewers"
          },
          "role": {
            "type": "string",
            "nullable": true,
            "description": "The viewer's effective role on this artifact, or null if they have none"
          }
        },
        "required": [
          "id",
          "name",
          "role"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/agents/{id}": {
      "delete": {
        "tags": [
          "Agents"
        ],
        "summary": "Delete an agent (Admin only).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "The agent was deleted."
          }
        }
      }
    },
...
    "/v1/me/connected-agents/{clientId}": {
      "delete": {
        "tags": [
          "Agents"
        ],
        "summary": "Revoke a connected agent's authorization.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "clientId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "The grant was revoked (idempotent)."
          }
        }
      }
    },
...
    "/v1/artifacts/{shortId}/presence": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "summary": "Presence heartbeat — keep this viewer on the artifact's roster.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "shortId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The current viewers on this artifact.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "viewers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Viewer"
                      }
                    }
                  },
                  "required": [
                    "viewers"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/artifacts/{shortId}/cursor": {
      "post": {
        "tags": [
          "Realtime"
        ],
        "summary": "Broadcast a live cursor frame to the artifact's other viewers.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "shortId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Fanned out (ephemeral; nothing stored)."
          }
        }
      }
    },
...
  }
}